- 27.09.17 CHG : signal/Functionals : added 'global' option to extract functionals across all dimensions : Johannes - 29.09.17 BUG : core/ioput/option : fixed small bug in parsing options from string : Johnny - 29.09.17 BUG : plugins/model/ClassifierT : fixed bug with stream merging : Johnny - 04.10.17 CHG : core/ioput/option : add read-only flag again : Johannes - 04.10.17 CHG : nova : model evaluation : Johannes - 05.10.17 BUG : model/Evaluation : now handles garbage samples : Johannes - 04.10.17 BUG : core/ioput/option : applied previous fix again : Johannes - 06.10.17 BUG : nova : loads local files if no server string is set : Johannes - 17.10.17 CHG : event/EventMonitor : option to show only last event : Johannes - 17.10.17 CHG : event/EventMonitor : options to set font name and size : Johannes - 17.10.17 BUG : core/ioput/option : correctly reads options with multiple values : Johannes - 18.10.17 CHG : ioput/XMLEventSender : added new field option 'isnew' : Johannes

+++ plugins/biosig/source/GSRFeatures.cpp
@@ -262,7 +262,7 @@ void GSRFeatures::transform( ITransformer::info info, ssi_stream_t &stream_in, s
 		if (maxima.size() > 0) avg_maxima /= ((float)maxima.size());
 		if (minima.size() > 0) avg_minima /= ((float) minima.size());
 		
-		f_p2pmv = std::fabsf(avg_maxima - avg_minima);
+        f_p2pmv = std::fabs(avg_maxima - avg_minima);
 
 	}
 	
@@ -528,7 +528,7 @@ float GSRFeatures::getStDev(std::vector<peakslopedrop> in, ssi_size_t attribute)
 
 		if (avg != 0){
 			var /= avg;
-			r = std::sqrtf(var);
+            r = std::sqrt(var);
 		}
 		else{
 			r = 0;
